pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pydns-scanner"
version = "2.0.3"
description = "A high-performance DNS scanner with TUI, SlipNet/Slipstream tunnel testing, DNS tunnel-compatibility scoring (NS/TXT/RND/DPI/EDNS0/NXD), real-time ISP/DNSSEC/security checks, and Iran CIDR bundle"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "xullexer" }
]
keywords = [
    "dns",
    "scanner",
    "tui",
    "terminal",
    "network",
    "async",
    "textual",
    "proxy",
    "slipstream",
    "cidr"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: Name Service (DNS)",
    "Topic :: System :: Networking",
    "Topic :: Utilities",
]
dependencies = [
    "textual>=0.47.0",
    "dnspython>=2.4.0",
    "httpx[socks]>=0.25.0",
    "loguru>=0.7.0",
    "pyperclip>=1.8.0",
]

[project.optional-dependencies]
full = [
    "google-re2>=1.0",
    "orjson>=3.9.0",
    "pyperclip>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/xullexer/PYDNS-Scanner"
Repository = "https://github.com/xullexer/PYDNS-Scanner"
Issues = "https://github.com/xullexer/PYDNS-Scanner/issues"

[project.scripts]
pydns-scanner = "python:main"

[tool.hatch.build.targets.sdist]
include = [
    "/python",
    "/README.md",
    "/LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["python"]

# Include CIDR data files in the package
[tool.hatch.build.targets.wheel.shared-data]
"python/iran-ipv4.cidrs" = "python/iran-ipv4.cidrs"
